backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
supabase==2.4.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
deepface==0.0.93
opencv-python-headless==4.10.0.84
numpy>=1.24.0,<2.0
Pillow>=10.0.0
python-dotenv==1.0.1
httpx>=0.27.0
aiofiles>=23.0.0
requests>=2.31.0
tf-keras>=2.16.0

